Add files via upload
I wrote all the code in the sections where you put comments asking me to write code. I also wrote comments explaining the logic and understanding behind my code.

There was one section of the code where you wrote comments asking to write code. It was the only section in which I didn't quite understand what to write. I left it empty. You'll see it in the code but just to make it easier for you, it was this section:


  // Make sure that PID does not reach more than let's say 60 positions so 90+60 = 150 or 90-60 = 30


I know we discussed this but I forgot the details of the discussion for this. I couldn't figure out how to implement this so I left it empty.