Add files via upload
This version has another function that I added: checkAltitude() 

This is going to check for the altitude of the drone to make sure it's not exceeding 1 meter. 

If it exceeds 1 meter => change motor speed so that drone drops in altitude. 

If it does not exceed 1 meter => leave motorSpeed as it is.